feat: Google Vision + GPT 연동 완료 (AI 스파 추천 시스템 연결)
- ImageCleanup 관련 로직 삭제 롤백
- 인증이 필요한 API에 @PreAuthorize("isAuthenticated()") 적용
- 추후 프론트에서 받아올 체크리스트 dto 추가

+++ src/main/java/com/example/oyl/repository/AiRecommendHistoryRepository.java
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 public interface AiRecommendHistoryRepository extends JpaRepository<AiRecommendHistory, Long> {
 
     // 하루 요청 제한 확인용!
-    int countByUserIdAndCreatedAtBetween(String userId, LocalDateTime start, LocalDateTime end);
+    long countByUserIdAndCreatedAtBetween(String userId, LocalDateTime start, LocalDateTime end);
 
 }
